Переделал логику репозитория и сервиса
