+++ src/main/java/lopatin/Repository.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 // Репозиторий для работы с dto
-public class Repository<T extends HumanDTO> {
+public class Repository<T extends Human> {
     /* Костыль для генерации случайных строк */
     private String stringGen() {
         String string = "randomletters";
@@ -16,36 +16,36 @@ public class Repository<T extends HumanDTO> {
     }
 
     // Генерация одного человека
-     HumanDTO get(HumanDTO dto) {
-        HumanDTO.Address address = new HumanDTO().new Address();
-        dto.setID((int) (Math.random() * 10));
-        dto.setName(stringGen());
-        dto.setBirthDate(new Date((long) (Math.random() * 100000)));
+     Human get(Human human) {
+        Human.Address address = new Human().new Address();
+        human.setID((int) (Math.random() * 10));
+        human.setName(stringGen());
+        human.setBirthDate(new Date((long) (Math.random() * 100000)));
         address.setCity(stringGen());
         address.setStreet(stringGen());
-        dto.setAddress(address);
-        return dto;
+        human.setAddress(address);
+        return human;
     }
 
     // Генерация i-го количества людей
-     List<HumanDTO> getAll(int i) {
-         Repository rep = new Repository();
-        ArrayList<HumanDTO> list = new ArrayList<>();
+     List<Human> getAll(int i) {
+        Repository rep = new Repository();
+        ArrayList<Human> list = new ArrayList<>();
         for (int j = 0; j < i; j++) {
-            HumanDTO dto = new HumanDTO();
-            list.add(rep.get(dto));
+            Human human = new Human();
+            list.add(rep.get(human));
         }
         return list;
     }
 
-    // Вывод инфы об одном dto
-     void save(HumanDTO dto) {
-        System.out.println(dto);
+    // Сохранение записи в БД
+     void save(Human human) {
+        System.out.println(human);
     }
 
-    // Вывод инфы о всех dto
-     void saveAll(List<HumanDTO> dtos) {
-        for (HumanDTO go : dtos) {
+    // Сохранение всех людей в БД
+     void saveAll(List<Human> humans) {
+        for (Human go : humans) {
             System.out.println(go);
         }
     }
