+++ src/main/java/lopatin/Service.java
@@ -1,52 +1,46 @@
 package lopatin;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
-public class Service<T extends Human> {
-    /* Костыль для генерации случайных строк */
-     private String stringGen() {
-        String string = "randomletters";
-        StringBuilder randString = new StringBuilder();
-        int count = (int) (Math.random() * 8 + 2);
-        for (int i = 0; i < count; i++)
-            randString.append(string.charAt((int) (Math.random() * string.length())));
-        return randString.toString();
+public class Service {
+
+    // Получение сущности из БД через ДТО
+    HumanDTO get(HumanDTO dto) {
+        Converter converter = new Converter();
+        Repository repository = new Repository();
+        Human human= converter.ConvertToHuman(dto);
+        repository.get(human);
+        dto = converter.ConvertToDTO(human);
+        return dto;
     }
 
-    // Генерация одного человека
-    Human get(Human human) {
-        Human.Address address = new Human(). new Address();
-        human.setID((int) (Math.random() * 10));
-        human.setName(stringGen());
-        human.setBirthDate(new Date((long) (Math.random() * 100000000)));
-        address.setCity(stringGen());
-        address.setStreet(stringGen());
-        human.setAddress(address);
-        return human;
-    }
-
-    // Генерация i-го количества людей
-     List<Human> getAll(int i) {
+    // Получение списка сущностей из БД через DTO
+     List<HumanDTO> getAll(int i) {
          Service service = new Service();
-        List<Human> list = new ArrayList<>();
+        List<HumanDTO> list = new ArrayList<>();
         for (int j = 0; j < i; j++) {
-            Human human = new Human();
-            list.add(service.get(human));
+            HumanDTO dto = new HumanDTO();
+            list.add(service.get(dto));
         }
         return list;
     }
 
-    // Вывод инфы об одном человеке
-     void save(Human human) {
-        System.out.println(human);
+    // Сохранение ДТО в БД путём конвертации в human
+     void save(HumanDTO dto) {
+        Converter converter = new Converter();
+        Repository repository = new Repository();
+        Human human=converter.ConvertToHuman(dto);
+        repository.save(human);
     }
 
-    // Вывод инфы о всех человеках
-     void saveAll(List<Human> persons) {
-        for (Human human : persons) {
-            System.out.println(human);
+    // Сохранение всех ДТО в БД
+     void saveAll(List<HumanDTO> dtos) {
+        for (HumanDTO dto : dtos) {
+            Converter converter = new Converter();
+            Repository repository = new Repository();
+            Human human=converter.ConvertToHuman(dto);
+            repository.save(human);
         }
     }
 }
